Remove methods from example apis.
