+++ examples/hello-world/src/test-api.js
@@ -31,27 +31,25 @@ const greetings = [
 // For demo purposes, track each request count and give an extra greeting each time.
 let requestCount = 0;
 
+export function fetchGreetings() {
+	return new Promise( ( resolve ) => {
+		requestCount++;
+		const valueCount = Math.min( requestCount, greetings.length );
+		const values = greetings.slice( 0, valueCount );
+		resolve( values );
+	} );
+}
+
 export default {
-	methods: {
-		get: ( endpointPath, params ) => { // eslint-disable-line no-unused-vars
-			return new Promise( ( resolve ) => {
-				requestCount++;
-				const valueCount = Math.min( requestCount, greetings.length );
-				const values = greetings.slice( 0, valueCount );
-				resolve( values );
-			} );
-		},
-	},
 	operations: {
-		read: ( { get } ) => ( resourceNames ) => {
+		read: ( resourceNames ) => {
 			const requests = [];
 			resourceNames.forEach( resourceName => {
 				if ( 'greetings' === resourceName ) {
-					const request = get( [ 'greetings' ] )
-						.then( data => {
-							const resources = { greetings: { data } };
-							return resources;
-						} );
+					const request = fetchGreetings().then( data => {
+						const resources = { greetings: { data } };
+						return resources;
+					} );
 					requests.push( request );
 				}
 			} );
